Specialize tokens.go for generating/validating tokens
main.go now is the example of how to use golin and
JWT implementation is "encapsulated" in an specialized
library
